server/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes -o requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
    #   starlette
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.1.8
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   loguru
    #   tqdm
cryptography==46.0.5
    # via google-auth
distro==1.9.0
    # via openai
exceptiongroup==1.3.1 ; python_full_version < '3.11'
    # via anyio
fastapi==0.128.8
    # via lp-generator-backend
google-api-core==2.29.0
    # via
    #   google-cloud-core
    #   google-cloud-firestore
google-auth==2.48.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-firestore
google-cloud-core==2.5.0
    # via google-cloud-firestore
google-cloud-firestore==2.23.0
    # via lp-generator-backend
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.78.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.78.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.13.0
    # via openai
loguru==0.7.3
    # via lp-generator-backend
openai==2.21.0
    # via lp-generator-backend
proto-plus==1.27.1
    # via
    #   google-api-core
    #   google-cloud-firestore
protobuf==6.33.5
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23 ; python_full_version < '3.10' and implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pycparser==3.0 ; python_full_version >= '3.10' and implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   lp-generator-backend
    #   openai
pydantic-core==2.41.5
    # via pydantic
requests==2.32.5
    # via google-api-core
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via openai
starlette==0.49.3
    # via fastapi
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   cryptography
    #   exceptiongroup
    #   fastapi
    #   grpcio
    #   openai
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
urllib3==2.6.3
    # via requests
uvicorn==0.39.0
    # via lp-generator-backend
win32-setctime==1.2.0 ; sys_platform == 'win32'
    # via loguru
